+++ open_mastr/mastr.py
@@ -1,4 +1,5 @@
 import os
+from sqlalchemy import inspect, create_engine
 
 # import xml dependencies
 from open_mastr.xml_download.utils_download_bulk import download_xml_Mastr
@@ -27,6 +28,7 @@ from open_mastr.utils.config import (
     create_data_dir,
     get_data_version_dir,
     get_project_home_dir,
+    get_output_dir,
     setup_logger,
 )
 import open_mastr.utils.orm as orm
@@ -34,6 +36,8 @@ import open_mastr.utils.orm as orm
 # import initialize_database dependencies
 from open_mastr.utils.helpers import (
     create_database_engine,
+    rename_table,
+    create_translated_database_engine,
 )
 
 # constants
@@ -45,33 +49,45 @@ log = setup_logger()
 
 class Mastr:
     """
-    :class:`.Mastr` is used to download the MaStR database and keep it up-to-date.
+    `Mastr` is used to download the MaStR database and keep it up-to-date.
 
-    A sql database is used to mirror the MaStR database. It can be filled with
+    A SQL database is used to mirror the MaStR database. It can be filled with
     data either from the MaStR-bulk download or from the MaStR-API.
 
-    .. code-block:: python
-
-       from open_mastr import Mastr
-
-       db = Mastr()
-       db.download()
-
     Parameters
-    ------------
-        engine: {'sqlite', sqlalchemy.engine.Engine}, optional
-            Defines the engine of the database where the MaStR is mirrored to. Default is 'sqlite'.
+    ----------
+    engine : {'sqlite', sqlalchemy.engine.Engine}, optional
+        Defines the engine of the database where the MaStR is mirrored to.
+        Default is 'sqlite'.
+    connect_to_translated_db: boolean, optional
+            Allows connection to an existing translated database. Default is 'False'.
+            Only for 'sqlite'-type engines.
+
+    !!! example
+        ```python
+        from open_mastr import Mastr
+
+        db = Mastr()
+        db.download()
+        ```
     """
 
-    def __init__(self, engine="sqlite") -> None:
+    def __init__(self, engine="sqlite", connect_to_translated_db=False) -> None:
 
         validate_parameter_format_for_mastr_init(engine)
 
+        self.output_dir = get_output_dir()
         self.home_directory = get_project_home_dir()
-        self._sqlite_folder_path = os.path.join(self.home_directory, "data", "sqlite")
+        self._sqlite_folder_path = os.path.join(self.output_dir, "data", "sqlite")
         os.makedirs(self._sqlite_folder_path, exist_ok=True)
 
-        self.engine = create_database_engine(engine, self.home_directory)
+        self.is_translated = connect_to_translated_db
+        if connect_to_translated_db:
+            self.engine = create_translated_database_engine(
+                engine, self._sqlite_folder_path
+            )
+        else:
+            self.engine = create_database_engine(engine, self._sqlite_folder_path)
 
         print(
             f"Data will be written to the following database: {self.engine.url}\n"
@@ -97,101 +113,86 @@ class Mastr:
     ) -> None:
         """
         Download the MaStR either via the bulk download or via the MaStR API and write it to a
-        sqlite database.
+        SQLite database.
 
         Parameters
-        ------------
-        method: {'API', 'bulk'}, optional
+        ----------
+        method : 'API' or 'bulk', optional
             Either "API" or "bulk". Determines whether the data is downloaded via the
             zipped bulk download or via the MaStR API. The latter requires an account
             from marktstammdatenregister.de,
             (see :ref:`Configuration <Configuration>`). Default to 'bulk'.
-        data: str or list or None, optional
+        data : str or list or None, optional
             Determines which types of data are written to the database. If None, all data is
-            used. If it is a list, possible entries are listed at the table
-            below with respect to the download method. Missing categories are
-            being developed. If only one data is of interest, this can be
-            given as a string. Default to None, where all data is included.
+            used. If it is a list, possible entries are listed below with respect to the download method. Missing categories are
+            being developed. If only one data is of interest, this can be given as a string. Default to None, where all data is included.
 
-            .. csv-table:: Values for data parameter
-                :header-rows: 1
-                :widths: 5 5 5
+            | Data                  | Bulk | API  |
+            |-----------------------|------|------|
+            | "wind"                | Yes  | Yes  |
+            | "solar"               | Yes  | Yes  |
+            | "biomass"             | Yes  | Yes  |
+            | "hydro"               | Yes  | Yes  |
+            | "gsgk"                | Yes  | Yes  |
+            | "combustion"          | Yes  | Yes  |
+            | "nuclear"             | Yes  | Yes  |
+            | "gas"                 | Yes  | Yes  |
+            | "storage"             | Yes  | Yes  |
+            | "electricity_consumer"| Yes  | No   |
+            | "location"            | Yes  | Yes  |
+            | "market"              | Yes  | No   |
+            | "grid"                | Yes  | No   |
+            | "balancing_area"      | Yes  | No   |
+            | "permit"              | Yes  | Yes  |
+            | "deleted_units"       | Yes  | No   |
+            | "retrofit_units"      | Yes  | No   |
+        date : None or `datetime.datetime` or str, optional
 
-                "Data", "Bulk", "API"
-                "wind", "Yes", "Yes"
-                "solar", "Yes", "Yes"
-                "biomass", "Yes", "Yes"
-                "hydro", "Yes", "Yes"
-                "gsgk", "Yes", "Yes"
-                "combustion", "Yes", "Yes"
-                "nuclear", "Yes", "Yes"
-                "gas", "Yes", "Yes"
-                "storage", "Yes", "Yes"
-                "electricity_consumer", "Yes", "No"
-                "location", "Yes", "Yes"
-                "market", "Yes", "No"
-                "grid", "Yes", "No"
-                "balancing_area", "Yes", "No"
-                "permit", "Yes", "Yes"
-                "deleted_units", "Yes", "No"
-                "retrofit_units", "Yes", "No"
+            | date                  | Bulk | API  |
+            |-----------------------|------|------|
+            | "today"                | latest files are downloaded from marktstammdatenregister.de  | -  |
+            | "20230101"      | If file from this date exists locally, it is used. Otherwise it throws an error (You can only receive todays data from the server)  | -   |
+            | "existing"               | Use latest downloaded zipped xml files, throws an error if the bulk download folder is empty  | -  |
+            | "latest"               | -  | Retrieve data that is newer than the newest data already in the table  |
+            | datetime.datetime(2020, 11, 27)      | -  | Retrieve data that is newer than this time stamp   |
+            | None      | set date="today"  | set date="latest"   |
 
-        date: None or :class:`datetime.datetime` or str, optional
-            For bulk method:
-
-            Either "today" or None if the newest data dump should be downloaded
-            rom the MaStR website. If an already downloaded dump should be used,
-            state the date of the download in the format
-            "yyyymmdd". Defaults to None.
-
-            For API method:
-
-            Specify backfill date from which on data is retrieved
-
-            Only data with modification time stamp greater than `date` is retrieved.
-
-            * `datetime.datetime(2020, 11, 27)`: Retrieve data which is newer than this
-              time stamp
-            * 'latest': Retrieve data which is newer than the newest data already in the table.
-
-              .. warning::
-
-                 Don't use "latest" in combination with "api_limit". This might lead to
-                 unexpected results.
-
-            * `None`: Complete backfill
-
-            Defaults to `None`.
-        bulk_cleansing: bool, optional
+            Default to `None`.
+        bulk_cleansing : bool, optional
             If True, data cleansing is applied after the download (which is recommended). Default
             to True.
-        api_processes: int or None or "max", optional
+        api_processes : int or None or "max", optional
             Number of parallel processes used to download additional data.
             Defaults to `None`. If set to "max", the maximum number of possible processes
             is used.
 
-            .. warning::
+            !!! warning
 
                 The implementation of parallel processes is currently under construction.
                 Please let the argument `api_processes` at the default value `None`.
-
-        api_limit: int or None, optional
-            Limit number of units that data is download for. Defaults to `None` which refers
+        api_limit : int or None, optional
+            Limit the number of units that data is downloaded for. Defaults to `None` which refers
             to query data for existing data requests, for example created by
-            :meth:`~.create_additional_data_requests`. Note: There is a limited number of
+            [`create_additional_data_requests`][open_mastr.soap_api.mirror.MaStRMirror.create_additional_data_requests]. Note: There is a limited number of
             requests you are allowed to have per day, so setting api_limit to a value is
             recommended.
-        api_chunksize: int or None, optional
+        api_chunksize : int or None, optional
             Data is downloaded and inserted into the database in chunks of `chunksize`.
             Defaults to 1000.
-        api_data_types: list or None, optional
-            Select type of additional data that should be retrieved. Choose from
-            "unit_data", "eeg_data", "kwk_data", "permit_data".  Defaults to all.
-        api_location_types: list or None, optional
-            Select type of location that should be retrieved. Choose from
+        api_data_types : list or None, optional
+            Select the type of additional data that should be retrieved. Choose from
+            "unit_data", "eeg_data", "kwk_data", "permit_data". Defaults to all.
+        api_location_types : list or None, optional
+            Select the type of location that should be retrieved. Choose from
             "location_elec_generation", "location_elec_consumption", "location_gas_generation",
             "location_gas_consumption". Defaults to all.
         """
+
+        if self.is_translated:
+            raise TypeError(
+                "You are currently connected to a translated database.\n"
+                "A translated database cannot be further processed."
+            )
 
         validate_parameter_format_for_download_method(
             method=method,
@@ -214,13 +215,12 @@ class Mastr:
             method, data, api_data_types, api_location_types, **kwargs
         )
 
-        date = transform_date_parameter(method, date, **kwargs)
+        date = transform_date_parameter(self, method, date, **kwargs)
 
         if method == "bulk":
-
             # Find the name of the zipped xml folder
             bulk_download_date = parse_date_string(date)
-            xml_folder_path = os.path.join(self.home_directory, "data", "xml_download")
+            xml_folder_path = os.path.join(self.output_dir, "data", "xml_download")
             os.makedirs(xml_folder_path, exist_ok=True)
             zipped_xml_file_path = os.path.join(
                 xml_folder_path,
@@ -296,7 +296,7 @@ class Mastr:
         ------------
         tables: None or list
             For exporting selected tables choose from:
-                ["wind","solar","biomass","hydro","gsgk","combustion","nuclear","storage",
+                ["wind", "solar", "biomass", "hydro", "gsgk", "combustion", "nuclear", "storage",
                 "balancing_area", "electricity_consumer", "gas_consumer", "gas_producer",
                 "gas_storage", "gas_storage_extended",
                 "grid_connections", "grids", "market_actors", "market_roles",
@@ -307,6 +307,13 @@ class Mastr:
         limit: None or int
             Limits the number of exported data rows.
         """
+
+        if self.is_translated:
+            raise TypeError(
+                "You are currently connected to a translated database.\n"
+                "A translated database cannot be used for the csv export."
+            )
+
         log.info("Starting csv-export")
 
         data_path = get_data_version_dir()
@@ -348,7 +355,6 @@ class Mastr:
 
         # Export technologies to csv
         for tech in technologies_to_export:
-
             db_query_to_csv(
                 db_query=create_db_query(tech=tech, limit=limit, engine=self.engine),
                 data_table=tech,
@@ -356,7 +362,6 @@ class Mastr:
             )
         # Export additional tables to csv
         for addit_table in additional_tables_to_export:
-
             db_query_to_csv(
                 db_query=create_db_query(
                     additional_table=addit_table, limit=limit, engine=self.engine
@@ -368,3 +373,59 @@ class Mastr:
         # FIXME: Currently metadata is only created for technology data, Fix in #386
         # Configure and save data package metadata file along with data
         # save_metadata(data=technologies_to_export, engine=self.engine)
+
+    def translate(self) -> None:
+        """
+        A database can be translated only once.
+
+        Deletes translated versions of the currently connected database.
+
+        Translates currently connected database,renames it with '-translated'
+        suffix and updates self.engine's path accordingly.
+
+        !!! example
+            ```python
+
+            from open_mastr import Mastr
+            import pandas as pd
+
+            db = Mastr()
+            db.download(data='biomass')
+            db.translate()
+
+            df = pd.read_sql(table='biomass_extended', con=db.engine)
+            print(df.head(10))
+            ```
+
+        """
+
+        if "sqlite" not in self.engine.dialect.name:
+            raise ValueError("engine has to be of type 'sqlite'")
+        if self.is_translated:
+            raise TypeError("The currently connected database is already translated.")
+
+        inspector = inspect(self.engine)
+        old_path = r"{}".format(self.engine.url.database)
+        new_path = old_path[:-3] + "-translated.db"
+
+        if os.path.exists(new_path):
+            try:
+                os.remove(new_path)
+            except Exception as e:
+                print(f"An error occurred: {e}")
+
+            print("Replacing previous version of the translated database...")
+
+        for table in inspector.get_table_names():
+            rename_table(table, inspector.get_columns(table), self.engine)
+
+        self.engine.dispose()
+
+        try:
+            os.rename(old_path, new_path)
+            print(f"Database '{old_path}' changed to '{new_path}'")
+        except Exception as e:
+            print(f"An error occurred: {e}")
+
+        self.engine = create_engine(f"sqlite:///{new_path}")
+        self.is_translated = True